リソースのアクセス制限を追加

+++ server/app/controllers/api/v1/companies_controller.rb
@@ -9,6 +9,7 @@ module Api
       # TOPページに表示する会社（ランダムに5社）
       def pickup
         companies = Company.order('RAND()').limit(5)
+        authorize companies
         render json: companies
       end
 
@@ -19,11 +20,14 @@ module Api
                     else
                       Company.page(params[:page] ||= 1).per(10).order(created_at: :desc)
                     end
+
+        authorize companies
         render json: companies
       end
 
       def search
         companies = Company.search(params[:keyword]).page(params[:page] ||= 1).per(10).order(created_at: :desc)
+        authorize companies
         render json: companies
       end
 
@@ -38,6 +42,7 @@ module Api
 
         # Companyの中から該当する会社を取り出す
         companies = Company.where(id: skill_id.ids)
+        authorize companies
 
         # 取得したデータの中からIdを取り出し、それを元にして会社情報を取得する
         render json: companies
@@ -45,6 +50,7 @@ module Api
 
       # 新規作成
       def create
+        authorize Company
         company = current_api_v1_user.companies.new(company_params)
         if company.save
           render json: company
@@ -55,11 +61,13 @@ module Api
 
       # 会社詳細情報
       def show
+        authorize @company
         render json: @company
       end
 
       # 会社情報更新
       def update
+        authorize @company
         if @company.update(company_params)
           render json: @company
         else
@@ -69,6 +77,7 @@ module Api
 
       # 会社情報削除
       def destroy
+        authorize @company
         if @company.destroy
           render json: @company
         else
