+++ server/app/controllers/api/v1/likes_controller.rb
@@ -6,10 +6,12 @@ module Api
       before_action :authenticate_api_v1_user!, only: %i[create destroy]
       def index
         likes = Like.where(company_id: params[:company_id])
+        authorize likes
         render json: likes
       end
 
       def create
+        authorize Like
         like = Like.new(user_id: current_api_v1_user.id, company_id: params[:company_id])
         if like.save
           render json: like
@@ -19,6 +21,7 @@ module Api
       end
 
       def destroy
+        authorize Like
         like = Like.find_by(user_id: current_api_v1_user.id, company_id: params[:company_id])
         if like.destroy
           render json: like
